requirements.txt
beautifulsoup4
pandas
pymongo
requests
joblib
tqdm
pillow
